Move write() to API module
